refactor: EmptyButton을 실제 용도에 맞는 이름으로 변경
버튼 모양이지만 클릭 동작을 쓰지 않는 컴포넌트이므로 Button 대신 Box라는 이름을 씀
