+++ src/pages/selectGroup/ui/index.ts
@@ -1,3 +1,3 @@
 export { default as CreateGroupLinkButton } from './CreateGroupLinkButton';
-export { default as EmptyButton } from './EmptyButton';
+export { default as EmptyBox } from './EmptyBox';
 export { default as GroupLinkButton } from './GroupLinkButton';
